2021/requirements.txt
black
isort
attrs
rich
